requirements.txt
amqp==5.3.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==4.2.1
blinker==1.8.2
Brotli==1.1.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Flask==3.0.3
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
MarkupSafe==2.1.5
mutagen==1.47.0
packaging==24.2
prompt_toolkit==3.0.48
pycryptodomex==3.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.6
yt-dlp==2024.10.22
zipp==3.20.2